Fix typos in RoleSeeder and enhance factory definitions; remove unused factory files and update total column type in orders table

+++ database/migrations/2025_03_20_121945_create_orders_table.php
@@ -17,7 +17,7 @@ return new class extends Migration {
             $table->foreignId('user_id')->constrained('users');
             $table->foreignId('plant_id')->constrained('plants');
             $table->integer('quantity');
-            $table->integer('total');
+            $table->decimal('total', 10, 2);
             $table->enum(
                 'status',
                 ['pending', 'success', 'annulled']
